ci(batch6): split streaming_impl at the 500-LoC cap, fix typo
`tool_parser/streaming_impl.rs` reached 511 lines, over the repo cap.
Move the `flush()` cluster verbatim into a new `streaming_flush.rs`
(334 / 188 lines); `safe_emit_len` becomes `pub(super)` since it is
now called across the module boundary. No behaviour change.

Also `Runtime-togglable` -> `Runtime-toggleable` in serve_args, which
the `typos` check rejects.
