+++ crates/spark-server/src/tool_parser.rs
@@ -486,6 +486,7 @@ mod qwen3_coder;
 mod qwen3_xml;
 mod streaming;
 mod streaming_emit;
+mod streaming_flush;
 mod streaming_impl;
 mod type_coerce;
 pub(crate) mod validation;
